requirements.txt
openai
aiogram
python-dotenv
pymongo
aiocron
langchain